+++ src/applications/reports/services/pdf/charts/use_of_vehicles/theoretical_use_of_vehicle.py
@@ -94,4 +94,4 @@ class TheoreticalUseOfVehicleChart(ChartGenerator):
         )
 
     def get_stats(self):
-        return np.array(self.reserved_hours, np.float), np.array(self.free_hours, np.float)
+        return np.array(self.reserved_hours, float), np.array(self.free_hours, float)
